03/03/2021 - 1500 - Working on improvements to the modules to get them running.

+++ pvt_model/pvt_system_model/__utils__.py
@@ -545,7 +545,7 @@ class CollectorParameters(OpticalLayerParameters):
     inner_pipe_diameter: float
     length: float
     mass_flow_rate: float
-    number_of_pipes: float
+    number_of_pipes: int
     outer_pipe_diameter: float
